moved the model function into model folder
